+++ quizx/src/hash_graph.rs
@@ -104,8 +104,8 @@ impl GraphLike for Graph {
         self.add_vertex_with_data(VData {
             ty,
             phase: Rational64::new(0, 1).into(),
-            qubit: 0,
-            row: 0,
+            qubit: 0.0,
+            row: 0.0,
         })
     }
 
@@ -212,19 +212,19 @@ impl GraphLike for Graph {
         Coord::new(d.qubit, d.row)
     }
 
-    fn set_qubit(&mut self, v: V, qubit: i32) {
+    fn set_qubit(&mut self, v: V, qubit: f64) {
         self.vdata.get_mut(&v).expect("Vertex not found").qubit = qubit;
     }
 
-    fn qubit(&self, v: V) -> i32 {
+    fn qubit(&self, v: V) -> f64 {
         self.vdata.get(&v).expect("Vertex not found").qubit
     }
 
-    fn set_row(&mut self, v: V, row: i32) {
+    fn set_row(&mut self, v: V, row: f64) {
         self.vdata.get_mut(&v).expect("Vertex not found").row = row;
     }
 
-    fn row(&self, v: V) -> i32 {
+    fn row(&self, v: V) -> f64 {
         self.vdata.get(&v).expect("Vertex not found").row
     }
